flight_tracker.py
import os
import sqlite3
import smtplib
from email.mime.text import MIMEText
from email.mime.multipart import MIMEMultipart
from datetime import datetime
from dotenv import load_dotenv
import time
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.webdriver.chrome.options import Options
from selenium.webdriver.chrome.service import Service
from webdriver_manager.chrome import ChromeDriverManager

# Load environment variables
load_dotenv()

class FlightTracker:
    def __init__(self):
        self.db_name = "flight_prices.db"
        self.init_database()
        
    def init_database(self):
        """Initialize SQLite database"""
        conn = sqlite3.connect(self.db_name)
        cursor = conn.cursor()
        
        # Create flights table
        cursor.execute('''
            CREATE TABLE IF NOT EXISTS flights (
                id INTEGER PRIMARY KEY AUTOINCREMENT,
                origin TEXT NOT NULL,
                destination TEXT NOT NULL,
                departure_date TEXT NOT NULL,
                email TEXT NOT NULL,
                target_price REAL,
                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
            )
        ''')
        
        # Create price_history table
        cursor.execute('''
            CREATE TABLE IF NOT EXISTS price_history (
                id INTEGER PRIMARY KEY AUTOINCREMENT,
                flight_id INTEGER,
                price REAL NOT NULL,
                currency TEXT DEFAULT 'INR',
                checked_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
                FOREIGN KEY (flight_id) REFERENCES flights (id)
            )
        ''')
        
        conn.commit()
        conn.close()
        
    def scrape_google_flights(self, origin, destination, departure_date):
        """Scrape flight prices from Google Flights"""
        print(f"🔍 Searching flights from {origin} to {destination} on {departure_date}...")
        
        try:
            # Setup Chrome options for headless browsing
            chrome_options = Options()
            chrome_options.add_argument("--headless")
            chrome_options.add_argument("--no-sandbox")
            chrome_options.add_argument("--disable-dev-shm-usage")
            chrome_options.add_argument("--disable-blink-features=AutomationControlled")
            chrome_options.add_argument("user-agent=Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36")
            
            # Initialize driver
            service = Service(ChromeDriverManager().install())
            driver = webdriver.Chrome(service=service, options=chrome_options)
            
            # Format date for URL (YYYY-MM-DD)
            date_formatted = departure_date
            
            # Build Google Flights URL
            url = f"https://www.google.com/travel/flights?q=Flights%20to%20{destination}%20from%20{origin}%20on%20{date_formatted}"
            
            print(f"📡 Fetching: {url}")
            driver.get(url)
            
            # Wait for prices to load
            time.sleep(5)
            
            # Try to find price elements
            prices = []
            try:
                price_elements = driver.find_elements(By.CSS_SELECTOR, "div[class*='YMlIz FpEdX'], span[class*='YMlIz FpEdX']")
                
                for element in price_elements[:3]:  # Get first 3 prices
                    price_text = element.text
                    if '₹' in price_text or 'INR' in price_text:
                        # Extract numeric value
                        price_numeric = ''.join(filter(lambda x: x.isdigit(), price_text))
                        if price_numeric:
                            prices.append(int(price_numeric))
            except Exception as e:
                print(f"⚠️ Could not find price elements: {e}")
            
            driver.quit()
            
            if prices:
                min_price = min(prices)
                print(f"✅ Found lowest price: ₹{min_price}")
                return min_price
            else:
                # Return demo price if scraping fails
                print("⚠️ Could not scrape real prices. Using demo price.")
                return 5500  # Demo price
                
        except Exception as e:
            print(f"❌ Error scraping flights: {e}")
            return 5500  # Demo price on error
    
    def add_flight(self, origin, destination, departure_date, email, target_price=None):
        """Add a flight to track"""
        conn = sqlite3.connect(self.db_name)
        cursor = conn.cursor()
        
        cursor.execute('''
            INSERT INTO flights (origin, destination, departure_date, email, target_price)
            VALUES (?, ?, ?, ?, ?)
        ''', (origin, destination, departure_date, email, target_price))
        
        flight_id = cursor.lastrowid
        conn.commit()
        conn.close()
        
        print(f"✅ Flight added! ID: {flight_id}")
        return flight_id
    
    def get_all_flights(self):
        """Get all tracked flights"""
        conn = sqlite3.connect(self.db_name)
        cursor = conn.cursor()
        
        cursor.execute('SELECT * FROM flights')
        flights = cursor.fetchall()
        
        conn.close()
        return flights
    
    def check_price(self, flight_id):
        """Check current price for a flight"""
        conn = sqlite3.connect(self.db_name)
        cursor = conn.cursor()
        
        # Get flight details
        cursor.execute('SELECT * FROM flights WHERE id = ?', (flight_id,))
        flight = cursor.fetchone()
        
        if not flight:
            print(f"❌ Flight ID {flight_id} not found")
            return None
        
        origin = flight[1]
        destination = flight[2]
        departure_date = flight[3]
        email = flight[4]
        target_price = flight[5]
        
        # Scrape current price
        current_price = self.scrape_google_flights(origin, destination, departure_date)
        
        if current_price:
            # Save to price history
            cursor.execute('''
                INSERT INTO price_history (flight_id, price)
                VALUES (?, ?)
            ''', (flight_id, current_price))
            conn.commit()
            
            # Get previous price
            cursor.execute('''
                SELECT price FROM price_history 
                WHERE flight_id = ? 
                ORDER BY checked_at DESC 
                LIMIT 1 OFFSET 1
            ''', (flight_id,))
            
            previous = cursor.fetchone()
            previous_price = previous[0] if previous else None
            
            # Check if price dropped
            if previous_price and current_price < previous_price:
                price_drop = previous_price - current_price
                print(f"🎉 PRICE DROP! ₹{price_drop} cheaper!")
                self.send_alert(email, origin, destination, departure_date, 
                              current_price, previous_price, price_drop)
            
            # Check if below target price
            if target_price and current_price <= target_price:
                print(f"🎯 TARGET PRICE REACHED! Current: ₹{current_price}, Target: ₹{target_price}")
                self.send_target_alert(email, origin, destination, departure_date, 
                                     current_price, target_price)
        
        conn.close()
        return current_price
    
    def send_alert(self, to_email, origin, destination, date, current_price, old_price, drop):
        """Send price drop alert email"""
        try:
            from_email = os.getenv('FROM_EMAIL')
            password = os.getenv('EMAIL_PASSWORD')
            
            msg = MIMEMultipart()
            msg['From'] = from_email
            msg['To'] = to_email
            msg['Subject'] = f'✈️ Flight Price Drop Alert: {origin} → {destination}'
            
            body = f"""
            <html>
            <body>
                <h2>🎉 Great News! Flight Price Dropped!</h2>
                <p><strong>Route:</strong> {origin} → {destination}</p>
                <p><strong>Date:</strong> {date}</p>
                <p><strong>Previous Price:</strong> ₹{old_price}</p>
                <p><strong>Current Price:</strong> ₹{current_price}</p>
                <p><strong>You Save:</strong> ₹{drop} 💰</p>
                <hr>
                <p>Book now before prices go up again!</p>
            </body>
            </html>
            """
            
            msg.attach(MIMEText(body, 'html'))
            
            server = smtplib.SMTP(os.getenv('SMTP_SERVER'), int(os.getenv('SMTP_PORT')))
            server.starttls()
            server.login(from_email, password)
            server.send_message(msg)
            server.quit()
            
            print(f"📧 Alert sent to {to_email}")
            
        except Exception as e:
            print(f"❌ Failed to send email: {e}")
    
    def send_target_alert(self, to_email, origin, destination, date, current_price, target_price):
        """Send target price reached alert"""
        try:
            from_email = os.getenv('FROM_EMAIL')
            password = os.getenv('EMAIL_PASSWORD')
            
            msg = MIMEMultipart()
            msg['From'] = from_email
            msg['To'] = to_email
            msg['Subject'] = f'🎯 Target Price Reached: {origin} → {destination}'
            
            body = f"""
            <html>
            <body>
                <h2>🎯 Your Target Price Has Been Reached!</h2>
                <p><strong>Route:</strong> {origin} → {destination}</p>
                <p><strong>Date:</strong> {date}</p>
                <p><strong>Your Target:</strong> ₹{target_price}</p>
                <p><strong>Current Price:</strong> ₹{current_price}</p>
                <hr>
                <p>Perfect time to book! 🎉</p>
            </body>
            </html>
            """
            
            msg.attach(MIMEText(body, 'html'))
            
            server = smtplib.SMTP(os.getenv('SMTP_SERVER'), int(os.getenv('SMTP_PORT')))
            server.starttls()
            server.login(from_email, password)
            server.send_message(msg)
            server.quit()
            
            print(f"📧 Target alert sent to {to_email}")
            
        except Exception as e:
            print(f"❌ Failed to send email: {e}")
    
    def get_price_history(self, flight_id):
        """Get price history for a flight"""
        conn = sqlite3.connect(self.db_name)
        cursor = conn.cursor()
        
        cursor.execute('''
            SELECT price, checked_at FROM price_history 
            WHERE flight_id = ? 
            ORDER BY checked_at DESC
        ''', (flight_id,))
        
        history = cursor.fetchall()
        conn.close()
        
        return history