requirements.txt
requests
selenium
pandas
python-dotenv
schedule
tabulate
webdriver-manager